pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "verifiers"
authors = [
    {name = "William Brown", email = "williambrown97@gmail.com"},
]
dynamic = ["version"]  # Let hatchling handle version
description = "Verifiers for reinforcement learning with LLMs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11,<3.13"
keywords = ["reinforcement-learning", "llm", "rl", "grpo", "verifiable-environments", "multi-turn"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "openai",
    "datasets",
    "pillow", # sorry, will :(
    "pydantic>=2.11.7",
]

[project.optional-dependencies]
all = [
    "ruff",
    "pre-commit",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "requests",
    "transformers",
    "accelerate>=1.4.0",
    "deepspeed",
    "peft",
    "wandb",
    "rich",
    "trl>=0.17.0",
    "vllm>=0.9.2",
    "liger-kernel>=0.5.10",
    "nest-asyncio>=1.6.0",
    "ipykernel",
    "ipywidgets",
    "math-verify>=0.8.0",
]

dev = [
    "ruff",
    "pre-commit",
    "requests",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "nest-asyncio>=1.6.0",
    "ipykernel",
    "ipywidgets",
]

docs = [
    "sphinx",
    "myst-parser",
    "sphinx-rtd-theme"
]

train = [
    "torch>=2.7.0",
    "transformers",
    "accelerate>=1.4.0",
    "requests",
    "peft",
    "wandb",
    "rich",
    "trl>=0.17.0",
    "vllm>=0.9.2",
    "liger-kernel>=0.5.10",
    "deepspeed",
]

envs = [
    "math-verify>=0.8.0",
    "requests",
    "duckduckgo-search",
    "brave-search",
]


[project.scripts]
vf-eval = "verifiers.scripts.eval:main"
vf-init = "verifiers.scripts.init:main"
vf-install = "verifiers.scripts.install:main"
vf-vllm = "verifiers.inference.vllm_server:main"

# Hatchling configuration
[tool.hatch.version]
path = "verifiers/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/verifiers",
    "/tests",
    "/README.md",
    "/LICENSE",
]
exclude = [
    "/wandb",
    "/scratch",
    "/configs",
    "/environments",
]

[tool.hatch.build.targets.wheel]
packages = ["verifiers"]

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
    "twine>=6.1.0",
]

[project.urls]
Homepage = "https://github.com/willccbb/verifiers"
Documentation = "https://github.com/willccbb/verifiers"
Repository = "https://github.com/willccbb/verifiers.git"
Issues = "https://github.com/willccbb/verifiers/issues"

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "-ra",
    "--quiet",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "asyncio: marks tests as async tests",
    "parsers: marks tests for parser components",
    "rubrics: marks tests for rubric components",
    "environments: marks tests for environment components",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::UserWarning:transformers.*",
]
asyncio_mode = "auto"
norecursedirs = [".git", ".tox", "dist", "build", "*.egg", "__pycache__"]

[tool.coverage.run]
source = ["verifiers"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
